+++ src/settings/_registerSettings.ts
@@ -15,6 +15,7 @@ import { registerMovementMultiplier } from "./movementMultiplier";
 import { registerMovementSpecificAnimationsSetting } from "./movementAnimations";
 import { registerMovementDataPathSetting } from "./movementPropertyPath";
 import { registerTokenMoveSpeedSetting } from "./tokenSpeed";
+import { registerTokenHudAppearanceSettings } from "./tokenHudAppearance";
 import { registerSystemIntegrationSettings } from "./systemIntegration";
 import { registerUncapExplorationSetting } from "./uncapExploration";
 import { registerWorldFont } from "./worldFont";
@@ -32,6 +33,7 @@ export function registerSettings() {
 	registerScaleJumpFactorSetting();
 	registerMovementSpecificAnimationsSetting();
 	registerTokenMoveSpeedSetting();
+	registerTokenHudAppearanceSettings();
 	registerAllowPathBeyondRangeSetting();
 	registerWorldFont();
 	registerAutoPathSetting();